Update tools\create-book.cmd
This change updates the `create-book` command to properly create a
unified Markdown file from the entire curriculum contents. To run this
tool, go to the top directory and run `tools\create-book.cmd`. The tool
will read the SUMMARY.md file, and from that create the unified book in
`build\intro-book-complete.md`.

From there, you can view the unified file with grip, and then print it,
or save it to PDF.

Note that the print version will _not_ page break before H1 headers. To
do this, you'll need to save as HTML, and then use the browser
development mode to add the CSS `page-break-before: always;` style. I
found that messing with the CSS cascade directly in the debugger was the
only way to do this; tweaking the HTML seemed to have no effect. I
believe that the GitHub markdown scripts may be clobbering various HTML
elements.